+++ src/features/stock-ticker/index.js
@@ -1,4 +1,4 @@
 export { StockTicker } from "./components/StockTicker";
 export { updateStockTickerAction } from "./redux/actions";
 export { stockTickerReducer } from "./redux/reducer";
-export { stockMiddleware } from "./redux/middleware";
+export { stockTickerMiddleware } from "./redux/middleware";
