+++ src/store/store.js
@@ -1,5 +1,4 @@
 import { createStore, applyMiddleware, compose } from "redux";
-import { stockMiddleware } from "../features/stock-ticker";
 import { chartMiddleware } from "../features/chart";
 import { combinedReducer } from "./reducer";
 import { socketService } from "../services";
@@ -8,16 +7,19 @@ import { keyStatsMiddleware } from "../features/key-stats";
 import { latestNewsMiddleware } from "../features/latest-news";
 import { topPeersMiddleware } from "../features/peers";
 import { headlineMiddleware } from "../features/headline";
+import { stockTickerMiddleware } from "../features/stock-ticker";
+import { stockMiddleware } from "../middleware";
 
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
 const middleware = [
-  stockMiddleware({ socketService }),
   chartMiddleware({ socketService }),
   companyMiddleware({ socketService }),
   keyStatsMiddleware({ socketService }),
   latestNewsMiddleware({ socketService }),
   topPeersMiddleware({ socketService }),
-  headlineMiddleware({ socketService })
+  headlineMiddleware({ socketService }),
+  stockTickerMiddleware({ socketService }),
+  stockMiddleware({ socketService })
 ];
 
 export const store = createStore(
